Code/requirements.txt
torch>=1.1.0
torchvision
numpy
rasterio
scikit-image
matplotlib
tqdm
pillow
pyyaml
#pip freeze > requirements.txt
